Move existing CLI tests to their own file
